# docs/openapi/swagger.yaml
openapi: 3.0.0
info:
  title: 'Descarga API'
  version: '0.1'
servers:
  -
    url: 'http://localhost:8000'
    description: Local
paths:
...
  '/subscribers/{id}':
    get:
      tags:
        - Subscriber
      summary: 'Subscriber Retrieve'
      responses:
        '200':
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriberFullOutput'
    put:
      tags:
        - Subscriber
      summary: 'Subscriber Update'
      requestBody:
        description: 'Input data format'
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubscriberUpdateInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriberFullOutput'
components:
  schemas:
...
    SubscriberUpdateInput:
      required:
        - firstName
        - lastName
      properties:
        firstName:
          example: Alef
        lastName:
          example: Castelo
      type: object
    SubscriberFullOutput:
      properties:
        id:
          type: integer
          example: Alef
        firstName:
          type: string
          example: Alef
        lastName:
          type: string
          example: Castelo
        email:
          type: string
          example: alefaraujocastelo@gmail.com
      type: object
...